requirements.txt
Flask
gunicorn
pycryptodome
flask_cors